+++ src/routes/HU_B3140W.tsx
@@ -1037,7 +1037,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1054,7 +1054,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1071,7 +1071,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1088,7 +1088,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
 
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
@@ -1106,7 +1106,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1123,7 +1123,7 @@ const HU_B3140W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
