+++ src/routes/MA_B7200W.tsx
@@ -1484,7 +1484,7 @@ const MA_B7200W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1501,7 +1501,7 @@ const MA_B7200W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1518,7 +1518,7 @@ const MA_B7200W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -1535,7 +1535,7 @@ const MA_B7200W: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
