+++ src/routes/SA_B2410_290W.tsx
@@ -913,7 +913,7 @@ const SA_B2410: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
@@ -945,7 +945,7 @@ const SA_B2410: React.FC = () => {
       props.field !== undefined ? (sum = item["total_" + props.field]) : ""
     );
 
-    var parts = parseInt(sum).toString().split(".");
+    var parts = parseFloat(sum).toString().split(".");
     return parts[0] != "NaN" ? (
       <td colSpan={props.colSpan} style={{ textAlign: "right" }}>
         {parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",") +
